Ssf. Temporal disaggregation

+++ demetra-toolkit/demetra-toolkit-core/demetra-toolkit-basic/src/main/java/jdplus/stats/likelihood/DiffuseLikelihoodStatistics.java
@@ -26,7 +26,7 @@ public class DiffuseLikelihoodStatistics {
         return logLikelihood + transformationAdjustment;
     }
 
-    public double getEffectiveObservationsCount() {
+    public int getEffectiveObservationsCount() {
         return observationsCount-diffuseCount;
     }
